Update start.sh
added if statements to the removing of x11 folders and creation of mods folder. this will help not creating a new dir over the old one possibly deleting the old Mods folder with active mods in it.

added an echo statement at line 8 to also inform the user that the persistentdata folder has changed to persistent_data. this was an issue i had that i have to investigate further upon entering a new world when update 1.3 was pushed.

+++ files/start.sh
@@ -5,6 +5,7 @@
 if [ -d '/mnt/foundry/server' ] || [ -d '/mnt/foundry/persistentdata' ]; then
     echo "Old docker volume setup found!"
     echo "Change your volume from /your/path/:/mnt/foundry/server too /your/path/:/home/foundry/server_files"
+    echo "Change your volume from /your/path/:/mnt/foundry/persistentdata too /your/path/:/home/foundry/persistent_data"
     echo "Check release notes 1.3 for more information!"
     echo "https://github.com/luxusburg/docker-foundry/releases"
     exit 1
@@ -79,14 +80,22 @@ else
         fi    
     fi
 fi
-echo " "
-echo "Cleaning possible X11 leftovers"
-echo " "
-rm /tmp/.X0-lock
-rm -r /tmp/*
 
-# add Mods folder for future use
-mkdir -p $server_files/Mods 2>/dev/null
+if [ -f /tmp/.X0-lock ]; then
+    echo "Removing /tmp/.X0-lock"
+    rm /tmp/.X0-lock
+fi
+
+if [ -d /tmp/ -a -n "$(ls -A /tmp)" ]; then
+    echo "Removing all files in /tmp directory"
+    rm -r /tmp/*
+    else
+        echo "No files to remove in /tmp directory"
+fi
+
+if [ ! -d "$server_files/Mods" ]; then
+    mkdir -p "$server_files/Mods" 2>/dev/null
+fi
 
 cd "$server_files"
 echo "Starting Foundry Dedicated Server"
